Revert "PSFR-1345 Add initial redis caching of apis (#364)"
This reverts commit 22c0bbf7e27853aeff78ab98027901cd79d075af.
