+++ src/main/resources/application-dev.yml
@@ -73,8 +73,3 @@ spring:
 
   feature-toggles:
     afine: true
-
-  data:
-    redis:
-      host: localhost
-      port: 6380
