+++ src/main/resources/application.yml
@@ -14,10 +14,7 @@ spring:
   mvc:
     async:
       request-timeout: 1m
-  data:
-    redis:
-      repositories:
-        enabled: false
+
   security:
     oauth2:
       resourceserver:
